fixed collision a bit by using floats

+++ Engine/Game.h
@@ -47,7 +47,7 @@ private:
 	/*  User Functions              */
 	void NewEnemy();
 	void DrawGameOver(int x, int y);
-	bool Collides(int ts);
+	bool Collides(float ts);
 	/********************************/
 private:
 	MainWindow& wnd;
